requirements.txt
asgi-redis==1.2.1
asgiref==1.1.2
asn1crypto==0.24.0
astroid==1.6.2
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
daphne==1.4.2
Django==2.0.2
django-cities==0.5.0.5
django-cors-headers==2.2.0
django-forms-bootstrap==3.1.0
django-jsonfield==1.0.1
django-model-utils==3.1.1
django-paypal==0.4.1
django-stripe-payments==2.0.0
django-webpack-loader==0.2.4
djangorestframework==3.7.7
djangorestframework-jsonapi==2.4.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
ecdsa==0.13
Fabric==1.10.2
future==0.16.0
geopy==1.11.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.5.5
mysqlclient==1.3.12
optional-django==0.3.0
paramiko==2.4.0
pbr==3.1.1
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
pycryptodome==3.5.1
PyJWT==1.6.1
pylint==1.8.3
PyNaCl==1.2.1
python-jose==2.0.2
pytz==2018.3
redis==2.10.6
requests==2.18.4
six==1.11.0
stevedore==1.28.0
stripe==1.79.1
swapper==1.1.0
tqdm==4.19.5
Twisted==17.9.0
txaio==2.9.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
webpack==6.0.0
wrapt==1.10.11
zope.interface==4.4.3
